modify and move helper-class
