--- app/SocialAccountService.php
@@ -1,37 +1,0 @@
-<?php
-
-namespace App;
-
-use Laravel\Socialite\Contracts\User as ProviderUser;
-
-class SocialAccountService
-{
-    public function findOrCreate(ProviderUser $providerUser, $provider)
-    {
-        $account = LinkedSocialAccount::where("provider_name", $provider)
-                   ->where("provider_id", $providerUser->getId())
-                   ->first();
-
-        if ($account) {
-            return $account->user;
-        } else {
-
-        $user = User::where("email", $providerUser->getEmail())->first();
-
-        if (! $user) {
-            $user = User::create([  
-                "email" => $providerUser->getEmail(),
-                "name"  => $providerUser->getName(),
-            ]);
-        }
-
-        $user->accounts()->create([
-            "provider_id"   => $providerUser->getId(),
-            "provider_name" => $provider,
-        ]);
-
-        return $user;
-
-        }
-    }
-}
