+++ composer.json
@@ -9,7 +9,7 @@
         "doctrine/dbal": "^2.9",
         "fideloper/proxy": "~4.0",
         "laravel/framework": "5.8.*",
-        "laravel/socialite": "^4.1",
+        "laravel/socialite": "^4.2",
         "laravel/tinker": "~1.0",
         "laravelcollective/html": "5.8.*",
         "league/flysystem-aws-s3-v3": "^1.0",
